Create abstract TestBaseSliders class
