--- osu.Game.Rulesets.Osu.Tests/TestCaseInput.cs
@@ -1,26 +1,0 @@
-// Copyright (c) 2007-2019 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
-
-using System;
-using System.Collections.Generic;
-using osu.Game.Rulesets.Osu.Objects;
-using osu.Game.Rulesets.Osu.Objects.Drawables;
-using osu.Game.Rulesets.Osu.Objects.Drawables.Pieces;
-using osu.Game.Tests.Visual;
-
-namespace osu.Game.Rulesets.Osu.Tests
-{
-    public class TestCaseInput : OsuTestCase
-    {
-        public override IReadOnlyList<Type> RequiredTypes => new[]
-        {
-            typeof(SliderBall),
-            typeof(SliderBody),
-            typeof(SliderTick),
-            typeof(DrawableSlider),
-            typeof(DrawableSliderTick),
-            typeof(DrawableRepeatPoint),
-            typeof(DrawableOsuHitObject)
-        };
-    }
-}
