+++ osu.Game.Rulesets.Osu.Tests/TestCaseSlider.cs
@@ -1,50 +1,23 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using System;
 using System.Collections.Generic;
+using osu.Game.Audio;
+using osu.Game.Rulesets.Osu.Objects;
+using osuTK;
+using osu.Game.Rulesets.Mods;
+using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Audio;
-using osu.Game.Beatmaps;
-using osu.Game.Beatmaps.ControlPoints;
-using osu.Game.Rulesets.Osu.Objects;
-using osu.Game.Rulesets.Osu.Objects.Drawables;
-using osu.Game.Tests.Visual;
-using osuTK;
-using osuTK.Graphics;
-using osu.Game.Rulesets.Mods;
-using System.Linq;
-using NUnit.Framework;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Objects;
-using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.Objects.Types;
-using osu.Game.Rulesets.Osu.Objects.Drawables.Pieces;
+
 
 namespace osu.Game.Rulesets.Osu.Tests
 {
     [TestFixture]
-    public class TestCaseSlider : OsuTestCase
+    public class TestCaseSlider : TestBaseSliders
     {
-        public override IReadOnlyList<Type> RequiredTypes => new[]
-        {
-            typeof(SliderBall),
-            typeof(SliderBody),
-            typeof(SliderTick),
-            typeof(DrawableSlider),
-            typeof(DrawableSliderTick),
-            typeof(DrawableRepeatPoint),
-            typeof(DrawableOsuHitObject)
-        };
-
-        private readonly Container content;
-        protected override Container<Drawable> Content => content;
-
-        private int depthIndex;
-        protected readonly List<Mod> Mods = new List<Mod>();
-
         public TestCaseSlider()
         {
             base.Content.Add(content = new OsuInputManager(new RulesetInfo { ID = 0 }));
@@ -98,10 +71,15 @@ namespace osu.Game.Rulesets.Osu.Tests
             AddStep("Distance Overflow", () => testDistanceOverflow());
             AddStep("Distance Overflow 1 Repeat", () => testDistanceOverflow(1));
         }
+        
+        protected Container content;
+        protected override Container<Drawable> Content => content;
+        
+        protected readonly List<Mod> Mods = new List<Mod>();
 
-        private void testSimpleBig(int repeats = 0) => createSlider(2, repeats: repeats);
+        private void testSimpleBig(int repeats = 0) => CreateSlider(2, repeats: repeats);
 
-        private void testSimpleBigLargeStackOffset(int repeats = 0) => createSlider(2, repeats: repeats, stackHeight: 10);
+        private void testSimpleBigLargeStackOffset(int repeats = 0) => CreateSlider(2, repeats: repeats, stackHeight: 10);
 
         private void testDistanceOverflow(int repeats = 0)
         {
@@ -116,43 +94,25 @@ namespace osu.Game.Rulesets.Osu.Tests
                     new Vector2(52, -34)
                 }, 700),
                 RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats),
+                NodeSamples = CreateEmptySamples(repeats),
                 StackHeight = 10
             };
 
-            addSlider(slider, 2, 2);
+            AddSlider(slider, 2, 2);
         }
 
-        private void testSimpleMedium(int repeats = 0) => createSlider(5, repeats: repeats);
+        private void testSimpleMedium(int repeats = 0) => CreateSlider(5, repeats: repeats);
 
-        private void testSimpleSmall(int repeats = 0) => createSlider(7, repeats: repeats);
+        private void testSimpleSmall(int repeats = 0) => CreateSlider(7, repeats: repeats);
 
-        private void testSlowSpeed() => createSlider(speedMultiplier: 0.5);
+        private void testSlowSpeed() => CreateSlider(speedMultiplier: 0.5);
 
-        private void testShortSlowSpeed(int repeats = 0) => createSlider(distance: 100, repeats: repeats, speedMultiplier: 0.5);
+        private void testShortSlowSpeed(int repeats = 0) => CreateSlider(distance: 100, repeats: repeats, speedMultiplier: 0.5);
 
-        private void testHighSpeed(int repeats = 0) => createSlider(repeats: repeats, speedMultiplier: 15);
+        private void testHighSpeed(int repeats = 0) => CreateSlider(repeats: repeats, speedMultiplier: 15);
 
-        private void testShortHighSpeed(int repeats = 0) => createSlider(distance: 100, repeats: repeats, speedMultiplier: 15);
+        private void testShortHighSpeed(int repeats = 0) => CreateSlider(distance: 100, repeats: repeats, speedMultiplier: 15);
 
-        private void createSlider(float circleSize = 2, float distance = 400, int repeats = 0, double speedMultiplier = 2, int stackHeight = 0)
-        {
-            var slider = new Slider
-            {
-                StartTime = Time.Current + 1000,
-                Position = new Vector2(-(distance / 2), 0),
-                Path = new SliderPath(PathType.PerfectCurve, new[]
-                {
-                    Vector2.Zero,
-                    new Vector2(distance, 0),
-                }, distance),
-                RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats),
-                StackHeight = stackHeight
-            };
-
-            addSlider(slider, circleSize, speedMultiplier);
-        }
 
         private void testPerfect(int repeats = 0)
         {
@@ -167,10 +127,10 @@ namespace osu.Game.Rulesets.Osu.Tests
                     new Vector2(400, 0)
                 }, 600),
                 RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats)
+                NodeSamples = CreateEmptySamples(repeats)
             };
 
-            addSlider(slider, 2, 3);
+            AddSlider(slider, 2, 3);
         }
 
         private void testLinear(int repeats = 0) => createLinear(repeats);
@@ -191,10 +151,10 @@ namespace osu.Game.Rulesets.Osu.Tests
                     new Vector2(430, 0)
                 }),
                 RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats)
+                NodeSamples = CreateEmptySamples(repeats)
             };
 
-            addSlider(slider, 2, 3);
+            AddSlider(slider, 2, 3);
         }
 
         private void testBezier(int repeats = 0) => createBezier(repeats);
@@ -214,10 +174,10 @@ namespace osu.Game.Rulesets.Osu.Tests
                     new Vector2(430, 0)
                 }),
                 RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats)
+                NodeSamples = CreateEmptySamples(repeats)
             };
 
-            addSlider(slider, 2, 3);
+            AddSlider(slider, 2, 3);
         }
 
         private void testLinearOverlapping(int repeats = 0) => createOverlapping(repeats);
@@ -238,10 +198,10 @@ namespace osu.Game.Rulesets.Osu.Tests
                     new Vector2(0, -200)
                 }),
                 RepeatCount = repeats,
-                NodeSamples = createEmptySamples(repeats)
+                NodeSamples = CreateEmptySamples(repeats)
             };
 
-            addSlider(slider, 2, 3);
+            AddSlider(slider, 2, 3);
         }
 
         private void testCatmull(int repeats = 0) => createCatmull(repeats);
@@ -267,61 +227,8 @@ namespace osu.Game.Rulesets.Osu.Tests
                 NodeSamples = repeatSamples
             };
 
-            addSlider(slider, 3, 1);
+            AddSlider(slider, 3, 1);
         }
 
-        private List<List<SampleInfo>> createEmptySamples(int repeats)
-        {
-            var repeatSamples = new List<List<SampleInfo>>();
-            for (int i = 0; i < repeats; i++)
-                repeatSamples.Add(new List<SampleInfo>());
-            return repeatSamples;
-        }
-
-        private void addSlider(Slider slider, float circleSize, double speedMultiplier)
-        {
-            var cpi = new ControlPointInfo();
-            cpi.DifficultyPoints.Add(new DifficultyControlPoint { SpeedMultiplier = speedMultiplier });
-
-            slider.ApplyDefaults(cpi, new BeatmapDifficulty { CircleSize = circleSize, SliderTickRate = 3 });
-
-            var drawable = new DrawableSlider(slider)
-            {
-                Anchor = Anchor.Centre,
-                Depth = depthIndex++
-            };
-
-            foreach (var mod in Mods.OfType<IApplicableToDrawableHitObjects>())
-                mod.ApplyToDrawableHitObjects(new[] { drawable });
-
-            drawable.OnNewResult += onNewResult;
-
-            Add(drawable);
-        }
-
-        private float judgementOffsetDirection = 1;
-        private void onNewResult(DrawableHitObject judgedObject, JudgementResult result)
-        {
-            var osuObject = judgedObject as DrawableOsuHitObject;
-            if (osuObject == null)
-                return;
-
-            OsuSpriteText text;
-            Add(text = new OsuSpriteText
-            {
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Text = result.IsHit ? "Hit!" : "Miss!",
-                Colour = result.IsHit ? Color4.Green : Color4.Red,
-                TextSize = 30,
-                Position = osuObject.HitObject.StackedEndPosition + judgementOffsetDirection * new Vector2(0, 45)
-            });
-
-            text.Delay(150)
-                .Then().FadeOut(200)
-                .Then().Expire();
-
-            judgementOffsetDirection *= -1;
-        }
     }
 }
